+++ src/features/albalist/components/AlbaCard.tsx
@@ -1,0 +1,109 @@
+'use client';
+
+import Chip from '@components/common/chip/Chip';
+import { AlbaItem } from '@features/albalist/mock/mockData';
+import { differenceInCalendarDays, format, isAfter } from 'date-fns';
+import Image from 'next/image';
+import { useEffect, useRef, useState } from 'react';
+
+import { cn } from '@/shared/lib/cn';
+
+import AlbaDropdown from './AlbaDrodown';
+
+interface Props {
+  item: AlbaItem;
+  onClick: () => void;
+}
+
+const AlbaCard = ({ item, onClick }: Props) => {
+  const start = new Date(item.recruitmentStartDate);
+  const end = new Date(item.recruitmentEndDate);
+  const dDay = differenceInCalendarDays(end, new Date());
+  const isRecruiting = isAfter(end, new Date());
+
+  const [open, setOpen] = useState(false);
+  const dropdownRef = useRef<HTMLDivElement | null>(null);
+
+  useEffect(() => {
+    const handleClickOutside = (e: MouseEvent) => {
+      if (
+        dropdownRef.current &&
+        !dropdownRef.current.contains(e.target as Node)
+      ) {
+        setOpen(false);
+      }
+    };
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => document.removeEventListener('mousedown', handleClickOutside);
+  }, []);
+
+  const dDayClass = cn(
+    'px-25 whitespace-nowrap lg:px-48',
+    dDay < 0 && 'text-gray-400',
+    dDay >= 0 && dDay <= 3 && 'text-error brightness-150 font-semibold',
+    dDay > 3 && 'text-gray-600 hover:text-gray-900'
+  );
+
+  return (
+    <div
+      className="w-360 cursor-pointer flex-col gap-8 rounded-xl border border-gray-200 p-24 transition-shadow hover:shadow-md lg:w-477"
+      onClick={onClick}
+    >
+      <div className="relative flex h-180 w-full justify-end overflow-hidden rounded-lg">
+        <Image
+          fill
+          alt="알바 이미지"
+          className="object-cover"
+          src={item.imageUrls?.[0] || '/images/logo.svg'}
+        />
+      </div>
+
+      <div className="relative mt-12 flex items-center gap-8 text-sm text-gray-600">
+        {item.isPublic && <Chip active label="공개" variant="filled" />}
+        <Chip
+          active
+          label={isRecruiting ? '모집 중' : '모집 완료'}
+          variant="filled"
+        />
+        <span className="ml-8 whitespace-nowrap text-black">
+          {format(start, 'yyyy.MM.dd')} ~ {format(end, 'yyyy.MM.dd')}
+        </span>
+
+        <div ref={dropdownRef} className="relative ml-auto">
+          <Image
+            alt="드롭다운 아이콘"
+            className="cursor-pointer"
+            height={24}
+            src="/icons/kebab-menu.svg"
+            width={24}
+            onClick={e => {
+              e.stopPropagation();
+              setOpen(prev => !prev);
+            }}
+          />
+          {open && <AlbaDropdown item={item} />}
+        </div>
+      </div>
+
+      <h3 className="mt-12 text-lg font-semibold text-black">{item.title}</h3>
+
+      <div className="mt-12 flex w-full justify-center rounded-lg bg-gray-25 py-6 text-xs text-gray-600 dark:bg-gray-100">
+        <span className="border-r border-gray-200 px-24 whitespace-nowrap hover:text-gray-900 lg:px-48">
+          지원자 <span>{item.applyCount}명</span>
+        </span>
+        <span className="border-r border-gray-200 px-24 whitespace-nowrap lg:px-48">
+          스크랩 <span>{item.scrapCount}명</span>
+        </span>
+        <span
+          className={cn(
+            `last:border-0 ${dDayClass} px-25 whitespace-nowrap lg:px-48`
+          )}
+        >
+          {dDay < 0 ? '마감 완료' : `마감 D-${dDay}`}
+        </span>
+      </div>
+    </div>
+  );
+};
+
+export default AlbaCard;
