+++ src/features/albalist/pages/ListPage.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import AlbaListPage from '../components/AlbaListPage';
+
+const ListPage = () => {
+  return <AlbaListPage />;
+};
+
+export default ListPage;
